Re-architect video-recap into an independent skill bundle (stateless)
Split the monolithic video-recap skill into 5 self-contained skills + a thin
orchestrator. Each skill ships its own lib.py (merged config+common, no
cross-skill imports); skills communicate only through JSON/MP4 artifacts in a
shared work_dir. Per the agreed direction (5 skills + orchestrator, "go simple"):
the cross-stage workflow_state/fingerprint/resume machinery is removed in favour
of a simple exists+mtime cache in video-understanding plus cheap downstream reruns.

Bundle:
- video-understanding — detect/extract/asr/vlm + brief building (was pipeline
  steps 1-4.1 + narration.py brief half) -> understanding index + writing brief
- video-script — narration writing rules (SKILL.md) + lint/validate (narration.py
  lint half) -> validated narration.json
- video-cut — edit.py -> edited_source.mp4 + narration_mapped.json
- video-voiceover — tts.py -> tts_segments + tts_meta.json
- video-assemble — assemble.py (mux/duck/subtitles) -> recap_<name>.mp4 + srt/ass
- video-recap — ~150-line orchestrator chaining the above (subprocess), pausing
  for the agent to write narration.json; carries doctor

narration.py (1121 lines) split along the brief/lint seam into video-understanding
(brief.py) and video-script (narration.py). The 1249-line pipeline.py god-object
(workflow_state, step fingerprints, cut-artifact currency) is gone.

Tests: reorganized into tests/<skill>/, run one isolated process per skill
(`make test`) since each skill ships its own lib. 96 behavioral tests pass; the
~38 dropped tests all covered the removed resume/caching/CLI-state machinery.

Verified: pyflakes clean, all entrypoints import, orchestrator subprocess wiring
resolves sibling skills, `recap.py --doctor` OK.
