--- skills/video-recap/references/prompt-templates.md
@@ -1,35 +1,0 @@
-# Prompt 模板
-
-> 本文件被 `load_prompt()` 按 `### NAME` 切片读取。当前 CLI 只读取 VLM 模板；解说词由 Agent 根据 `agent_narration_brief.md` 撰写。
-
-## 目录
-
-| 模板名 | 用途 | 使用阶段 |
-|--------|------|----------|
-| VLM_DEPTH_PROMPT | 场景画面深度分析 | VLM 分析 |
-
-### VLM_DEPTH_PROMPT
-仔细观察这些视频帧，分析这个场景。分三部分输出：
-
-【描述】
-不超过80字，描述画面中正在发生什么（场景、人物、动作、表情、文字、事件）。
-
-【帧标签】
-逐帧列出每帧可见的人物动作和物品关系。每帧一行，格式：
-时间点 | 该帧可见的关键动作（1-2个完整短句，包含谁在做什么、物品如何被使用，如有字幕/文字也一并列出）
-示例：
-5.0s | 男子闯入房间, 头发蓬乱表情紧张
-10.0s | 男子俯身盯着床上男孩, 男孩睁眼惊醒
-15.0s | 男孩从床上坐起双手抱住头上的瓷枕准备防御
-
-【深层分析】
-不超过120字，分析：
-1. 角色的真实情绪或心理状态（通过表情、肢体语言、眼神判断）
-2. 人物之间的关系动态（亲密、紧张、疏离、试探等）
-3. 对话背后的潜台词（他们没说出口但显而易见的东西）
-
-要求：
-- 必须基于画面可见的证据推断，不做无根据的猜测
-- 不要用"象征""隐喻"等文学词汇
-- 用口语化表达："他明显是在试探她的反应"
-- 如果画面信息不足，直接说"信息不足"
