+++ tests/assemble/test_assemble_integration.py
@@ -1,3 +1,6 @@
+import sys
+from pathlib import Path
+sys.path.insert(0, str(Path(__file__).resolve().parents[2] / 'skills' / 'video-assemble' / 'scripts'))
 """Integration characterization tests for the ffmpeg render path.
 
 These run real ffmpeg/ffprobe to lock the behavior of assemble_video (ducking
@@ -10,10 +13,9 @@ from pathlib import Path
 
 import pytest
 
-sys.path.insert(0, str(Path(__file__).resolve().parents[1] / 'skills' / 'video-recap' / 'scripts'))
 
 from assemble import assemble_video  # noqa: E402
-from config import CONFIG  # noqa: E402
+from lib import CONFIG  # noqa: E402
 
 _HAVE_FFMPEG = bool(shutil.which("ffmpeg") and shutil.which("ffprobe"))
 pytestmark = pytest.mark.skipif(not _HAVE_FFMPEG, reason="ffmpeg/ffprobe not available")
